examples/integrations/requirements.txt
# Framework Integration Examples - Dependencies
# Install with: pip install -r requirements.txt

# Core package with all integrations
prompt-manager[all]

# Optional: Environment variable management
python-dotenv>=1.0.0

# Framework SDKs (already included in prompt-manager[all], listed for reference)
# openai>=1.57.0
# anthropic>=0.42.0
# langchain-core>=0.3.0
# litellm>=1.53.0

# Optional: Additional LangChain integrations
# langchain-openai>=0.2.0
# langchain-anthropic>=0.2.0
